Traktor: Fixed shortcut refactorization in Win32 UI.
[git-p4: depot-paths = "//depot/traktor/": change = 1198]

+++ code/Ui/Win32/UtilitiesWin32.cpp
@@ -8,7 +8,7 @@ namespace traktor
 
 VirtualKey translateKeyCode(uint32_t keyCode)
 {
-	const struct { NSUInteger keyCode; VirtualKey vkey; } c_translateTable[] =
+	const struct { uint32_t keyCode; VirtualKey vkey; } c_translateTable[] =
 	{
 		{ VK_TAB, VkTab },
 		{ VK_CLEAR, VkClear },
